cloud-functions/ingest-raw-data/requirements.txt
BS4
requests
xlrd
google-cloud-storage